QPID-8581: Fixes after review Moved query engine configuration from HttpPort to HttpManagementConfiguration

+++ broker-plugins/broker-query-engine/src/main/java/org/apache/qpid/server/query/engine/parsing/converter/NumberConverter.java
@@ -64,7 +64,7 @@ public final class NumberConverter
             .put(Long.class, arg -> arg)
             .put(Number.class, arg -> ((Number) arg).longValue())
             .put(Short.class, arg -> ((Short) arg).longValue())
-            .put(String.class, arg -> Long.parseLong((String)arg))
+            .put(String.class, arg -> ((Double) Double.parseDouble((String)arg)).longValue())
             .build()
         ).build();
